requirements.txt
torch
open-clip-torch
Pillow
python-dotenv
streamlit
requests
